pfa-fastAPI/requirements.txt
fastapi
uvicorn
pandas
numpy
tabulate
eto
requests
python-dotenv
openpyxl
mysql-connector-python